게시판 서비스 - (기능) 회원 삭제(탈퇴), 게시글 삭제, 게시글 조회수 기능 구현

+++ src/main/java/project/boardService/service/PostService.java
@@ -1,6 +1,7 @@
 package project.boardService.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.boardService.dto.PostDto;
@@ -9,6 +10,7 @@ import project.boardService.entity.Comment;
 import project.boardService.entity.Member;
 import project.boardService.entity.Post;
 import project.boardService.exception.DataNotFoundException;
+import project.boardService.exception.UnauthorizedAccessException;
 import project.boardService.repository.MemberRepository;
 import project.boardService.repository.PostRepository;
 
@@ -73,10 +75,16 @@ public class PostService {
 
 
     //post 검색 (findPostById) 별도 반환 DTO
+    @Transactional //조회수가 증가되어야 해서, 트랜젝셔널이 필요!
     public PostDto findByPostDtoById(Long postId) {
         Optional<Post> findPost = postRepository.findPostById(postId);
         if (findPost.isPresent()) {
             Post post = findPost.get();
+
+            //조회수 증가
+            postRepository.incrementViewCount(postId);
+
+            //DTO 반환
             return new PostDto(post.getId(), post.getTitle(), post.getContent(), post.getWriter().getName(),
                     post.getView(), post.getCreatedDateTime(), post.getModifiedDateTime());
         } else {
@@ -84,12 +92,28 @@ public class PostService {
         }
     }
 
-    //post 삭제 (deletePost)
+    //post 삭제 (deletePost) + 삭제 조건
     @Transactional
-    public Long deletePost(Post post) {
-        Long deleteId = post.getId();
+    public Long deletePost(Long postId) {
+        //1. Post 찾기
+        Optional<Post> findPost = postRepository.findPostById(postId);
+
+        //*예외처리1 : Post 검색 실패
+        if (!findPost.isPresent()) {
+            throw new DataNotFoundException("게시글을 찾을 수 없습니다.");
+        }
+
+        Post post = findPost.get();
+        //로그인된 사용자명
+        String loggedName = SecurityContextHolder.getContext().getAuthentication().getName();
+        //*예외처리2 : 작성자 != 로그인 회원
+        if (!post.getWriter().getName().equals(loggedName)) {
+            throw new UnauthorizedAccessException("이 댓글을 삭제할 권한이 없습니다.");
+        }
+
+        //post 삭제
         postRepository.deletePost(post);
-        return deleteId;
+        return postId;
     }
 
     //post 수정 (updatePost)
@@ -113,15 +137,11 @@ public class PostService {
         //1. Member 찾기(id 기반)
         Optional<Member> findMember = memberRepository.findById(memberId);
         if (findMember.isPresent()) {
-            List<Post> posts = postRepository.findPostByMember(findMember.get());
-            if (!posts.isEmpty()) {
-                return posts;
-            } else {
-                throw new DataNotFoundException(findMember.get().getName() + "이(가) 작성한 게시글이 없습니다.");
-            }
+            return postRepository.findPostByMember(findMember.get());
         } else {
             throw new DataNotFoundException("존재하지 않는 회원 입니다.");
         }
     }
+
 
 }
